+++ Sources/Navigator/EPUB/EPUBNavigatorViewModel.swift
@@ -159,7 +159,7 @@ final class EPUBNavigatorViewModel: Loggable {
         self.config = config
         editingActions = EditingActionsController(
             actions: config.editingActions,
-            rights: publication.rights
+            publication: publication
         )
         self.httpServer = httpServer
         self.publicationEndpoint = publicationEndpoint