+++ Sources/Navigator/PDF/PDFNavigatorViewController.swift
@@ -105,7 +105,10 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
         self.publicationEndpoint = publicationEndpoint
         publicationBaseURL = URL(string: baseURL.absoluteString.addingSuffix("/"))!
         self.config = config
-        editingActions = EditingActionsController(actions: config.editingActions, rights: publication.rights)
+        editingActions = EditingActionsController(
+            actions: config.editingActions,
+            publication: publication
+        )
 
         settings = PDFSettings(
             preferences: config.preferences,
@@ -135,7 +138,7 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
         publicationEndpoint = nil
         publicationBaseURL = URL(string: baseURL.absoluteString.addingSuffix("/"))!
         config = Configuration(editingActions: editingActions)
-        self.editingActions = EditingActionsController(actions: editingActions, rights: publication.rights)
+        self.editingActions = EditingActionsController(actions: editingActions, publication: publication)
 
         settings = PDFSettings(
             preferences: config.preferences,
@@ -175,7 +178,7 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
         self.publicationEndpoint = publicationEndpoint
         self.publicationBaseURL = URL(string: publicationBaseURL.absoluteString.addingSuffix("/"))!
         self.config = config
-        editingActions = EditingActionsController(actions: config.editingActions, rights: publication.rights)
+        editingActions = EditingActionsController(actions: config.editingActions, publication: publication)
 
         settings = PDFSettings(
             preferences: config.preferences,
@@ -205,8 +208,6 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
         super.viewDidLoad()
 
         resetPDFView(at: initialLocation)
-
-        editingActions.updateSharedMenuController()
     }
 
     override open func viewWillAppear(_ animated: Bool) {
@@ -541,9 +542,10 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
 
     @objc func selectionDidChange(_ note: Notification) {
         guard
+            ensureSelectionIsAllowed(),
             let pdfView = pdfView,
-            let locator = currentLocation,
             let selection = pdfView.currentSelection,
+            let locator = currentLocation,
             let text = selection.string,
             let page = selection.pages.first
         else {
@@ -558,15 +560,29 @@ open class PDFNavigatorViewController: UIViewController, VisualNavigator, Select
                 .insetBy(dx: -8, dy: -8)
         )
     }
-
-    @objc private func shareSelection(_ sender: Any?) {
-        guard
-            let pdfView = pdfView,
-            let shareViewController = editingActions.makeShareViewController(from: pdfView)
-        else {
-            return
+    
+    /// From iOS 13 to 15, the Share menu action is impossible to remove without
+    /// resorting to complex method swizzling in the subviews of ``PDFView``.
+    /// (https://stackoverflow.com/a/61361294)
+    ///
+    /// To prevent users from copying the text, we simply disable all text
+    /// selection in this case.
+    private func ensureSelectionIsAllowed() -> Bool {
+        guard !editingActions.canCopy else {
+            return true
         }
-        present(shareViewController, animated: true)
+
+        if #available(iOS 13, *) {
+            if #available(iOS 16, *) {
+                // Do nothing, as the issue is solved since iOS 16.
+            } else {
+                if let pdfView = pdfView, pdfView.currentSelection != nil {
+                    pdfView.clearSelection()
+                }
+                return false
+            }
+        }
+        return true
     }
 
     // MARK: - Navigator