+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file. Take a look
 
 **Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with caution.
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Fixed
+
+#### Navigator
+
+* From iOS 13 to 15, PDF text selection is disabled on protected publications disabling the **Copy** editing action.
+* The **Share** editing action is disabled for any protected publication.
+
 
 ## [2.6.1]